GenAI/Projects/Banking-Projects/Bank-guidelines-rbi-chatbot/app/requirements.txt
fastapi
uvicorn
langchain
openai
redis
python-dotenv
pdfplumber
pandas
chromadb
